tests/wpt/web-platform-tests/offscreen-canvas/line-styles/2d.line.width.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.line.width.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.line.width.valid</h1>
<p class="desc">Setting lineWidth to valid values works</p>


<script>
var t = async_test("Setting lineWidth to valid values works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.lineWidth = 1.5;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");
ctx.lineWidth = "1e1";
_assertSame(ctx.lineWidth, 10, "ctx.lineWidth", "10");
ctx.lineWidth = 1/1024;
_assertSame(ctx.lineWidth, 1/1024, "ctx.lineWidth", "1/1024");
ctx.lineWidth = 1000;
_assertSame(ctx.lineWidth, 1000, "ctx.lineWidth", "1000");
t.done();

});
</script>
